Fix bug performing uninitiated operation
When the equal button is pressed and a result
is generated, pressing an operator performs another
operation prior, as part of the handling from directly
pressing operators after entering numbers instead of equal
button. This fix prevents this behavior.

It also prevents successive presses of the equal
button from performing any action.